@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
 using Futbal.Mng.Domain.Event;
 using Futbal.Mng.Domain.UserManagement;
 using Futbal.Mng.Infrastructure.DTO;
@@ -12,12 +14,15 @@ namespace Futbal.Mng.Infrastructure.UserManagement.Services
     {
         private readonly IUserRepository _userRepository;
         private readonly IGameRepository _gameRepository;
+        private readonly IMapper _mapper;
 
         public UserService(IUserRepository userRepository,
-        IGameRepository gameRepository)
+        IGameRepository gameRepository,
+        IMapper mapper)
         {
             _userRepository = userRepository;
             _gameRepository = gameRepository;
+            _mapper = mapper;
         }
         public async Task<GameDetailsDto> GetAsync(Guid id)
         {
@@ -30,6 +35,18 @@ namespace Futbal.Mng.Infrastructure.UserManagement.Services
         {
             var user = new User(userDto.Username, userDto.Password, userDto.Email);
             await _userRepository.Add(user);
+        }
+
+        public async Task<IEnumerable<UserDto>> GetAllUsers()
+        {
+            var users = await _userRepository.GetUsers();
+
+            return users.Select(user => 
+            new UserDto
+                {
+                    Id = user.Id,
+                    Username = user.Username
+                });
         }
     }
 }
